Stop redisCheckSocketError from more than checking
redisCheckSocketError should only CheckSocketError and not
close any misbehaving sockets.  If CheckSocketError discovers
a problem, the caller will discover the contest is in ERR
and will start destroying the context (which involves
finalizing all callbacks (which may still be using
fd for something, so we should not close fd until all
callbacks have been told we are failing) and eventually
close the fd in redisFree() immediately before the
context is released).