Add reconnect logic to remaining connect functions
This captures the relevant details necessary for reconnecting on unix
sockets and bind connections. reconnect has been updated to do the right
thing based on the information provided in the context. An error will
be set on the context if there is not enough information to reconnect.

There were also a few indentation issues that needed to be cleaned
up. These were addressed in this commit as well.