avoid name clash with system specific predefined macro 'unix'
It is common to find 'unix' defined on Unix systems before c99. While we
can force that hiredis itself is compiled with c99, we cannot really
enforce that all projects using the hiredis headers are compiled with
c99.

So avoid the use of 'unix' in the public headers.

The problem was introduced with commit d9e0b0f6 (Implement a reconnect
method for the client context, 2015-04-16)

See also:
http://stackoverflow.com/questions/3770322/is-unix-restricted-keyword-in-c

This fixes #321