Always resolve v4 and v6 addresses
Otherwise the code will never work on v6-only hosts, when the domain
resolves to v4 addresses.

The latency is negligible. The order of resolved addresses is already a
good hint on which connectivity is available.
If IPv6 is returned, but no such connectivity is there, it's a system
configuration mistake.

Additionally, repetitive re-connections to Redis should be avoided
anyway.