Remove nested depth limitation.
This commit removes the nested multi-bulk depth limitation of 7.
We do this by switching to pointer to pointer indirection and
growing the stack in chunks when needed.

Initial benchmarking indicates the worse cache locality seems to be
offset by the reduction in overall redisReader size, resulting in
nearly identical parsing performance to the exiting master branch.

See: #794, #421