Patch Hiredis v1.0.0 to use Redis' SDS library.
Hiredis needs to run in Windows so has a fair bit of logic to handle
that environment.  This isn't a requirement when packaging it in Redis,
and to avoid having to maintain two copies of virtually the same library
in memory, just use the one from Redis.

There are a few places where Hiredis is checking for NULL return values
from allocations, because we try to gracefully handle those situations.
This isn't needed in Redis as it just hard aborts under such
circumstances.