Added REDIS_NO_AUTO_FREE_REPLIES flag
The new flag can be set on the redisOptions struct and given to the async connection.
If set, the new flag makes sure the replies given to the async callback will not be
freed by hiredis. Instead, the reply ownership is given to the callback and it's the user's
responsibility to free it when he finished using it.