Implement RESET in async API: Cancel subscriptions and monitor mode
Additionally, accept commands in monitor mode. (For example the RESET
command, but also other commands.)

Apart from being useful but itself, this change makes the async API's
reply queue stay in sync (mapping each reply to the callback given
when the command was sent) when bombed with random commands (fuzzing).