Remove `format_command_fuzzer.c`
The fuzzer can't surface any meaningful defects due to the nature of
variadic functions in C and how libfuzzer works.

The fuzzer just generates random strings that are passed into
`redisFormatCommand` with no subsequent arguments, which is a violation
of the contract for the function.

`redisFormatCommand("%b");` will result in undefined behavior in the
same way `printf("%s");` will, but it's not really a bug.

Fixes #1293, #1281